+++ src/main/resources/changelog.md
@@ -1,6 +1,9 @@
 ##0.5.7
 * Improve the removal of snow.
 * If you notice lag, feel free to turn off "Remove snow faster" config option
+* Validate some things before syncing player stats
+* Fix the beetroot state handler
+* Fix melons and pumpkins being unsellable
 
 ##0.5.6
 * Save the mine portal locations to the dimension, not level.dat